Create a separate file for each module in the modules directory
