+++ modules/js/bookCollection.js
@@ -1,20 +1,26 @@
-import Book from './book.js';
+import { Book } from './book.js';
+import { LocalStorage } from './localStorage.js';
 
-export default class BookCollection {
+export class BookCollection {
   constructor() {
-    this.books = JSON.parse(localStorage.getItem('books')) || [];
+    this.books = LocalStorage.getBooks() || [];
+  }
+
+  init() {
+    this.displayBooks();
+    this.setupForm();
   }
 
   addBook(title, author) {
     const book = new Book(title, author);
 
     this.books.push(book);
-    this.saveToLocalStorage();
+    LocalStorage.saveBooks(this.books);
   }
 
   removeBook(title) {
     this.books = this.books.filter((book) => book.title !== title);
-    this.saveToLocalStorage();
+    LocalStorage.saveBooks(this.books);
   }
 
   displayBooks() {
@@ -38,7 +44,29 @@ export default class BookCollection {
     });
   }
 
-  saveToLocalStorage() {
-    localStorage.setItem('books', JSON.stringify(this.books));
+  setupForm() {
+    const form = document.querySelector('.form');
+
+    form.addEventListener('submit', (event) => {
+      event.preventDefault();
+
+      const titleInput = document.getElementById('title');
+      const authorInput = document.getElementById('author');
+
+      this.addBook(titleInput.value, authorInput.value);
+      this.displayBooks();
+
+      titleInput.value = '';
+      authorInput.value = '';
+    });
+
+    document.getElementById('book-lists').addEventListener('click', (event) => {
+      if (event.target.classList.contains('remove-button')) {
+        const { title } = event.target.dataset;
+
+        this.removeBook(title);
+        this.displayBooks();
+      }
+    });
   }
 }
